Move Dirs classes into a dedicated subdirectory for better organisation
