+++ src/Config/Dirs/TrailingGlobDirs.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Config;
+namespace Haspadar\Piqule\Config\Dirs;
 
 use Override;
 
 /**
- * Directories suffixed with / to match directory entries explicitly
+ * Directories suffixed with /** for recursive glob exclusion patterns
  */
-final readonly class TrailingSlashDirs implements Dirs
+final readonly class TrailingGlobDirs implements Dirs
 {
     /** @param list<string> $dirs */
     public function __construct(private array $dirs) {}
@@ -18,6 +18,6 @@ final readonly class TrailingSlashDirs implements Dirs
     #[Override]
     public function toList(): array
     {
-        return array_map(fn(string $dir): string => $dir . '/', $this->dirs);
+        return array_map(fn(string $dir): string => $dir . '/**', $this->dirs);
     }
 }
