+++ tests/Unit/Config/Dirs/ProjectDirsTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Tests\Unit\Config;
+namespace Haspadar\Piqule\Tests\Unit\Config\Dirs;
 
-use Haspadar\Piqule\Config\ProjectDirs;
+use Haspadar\Piqule\Config\Dirs\ProjectDirs;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
