+++ tests/Unit/Config/Dirs/TrailingGlobDirsTest.php
@@ -1,0 +1,32 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Haspadar\Piqule\Tests\Unit\Config\Dirs;
+
+use Haspadar\Piqule\Config\Dirs\TrailingGlobDirs;
+use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\TestCase;
+
+final class TrailingGlobDirsTest extends TestCase
+{
+    #[Test]
+    public function returnsEmptyListWhenNoDirs(): void
+    {
+        self::assertSame(
+            [],
+            (new TrailingGlobDirs([]))->toList(),
+            'TrailingGlobDirs must return an empty list when constructed with no directories',
+        );
+    }
+
+    #[Test]
+    public function suffixesEachDirWithRecursiveGlob(): void
+    {
+        self::assertSame(
+            ['vendor/**', '.git/**'],
+            (new TrailingGlobDirs(['vendor', '.git']))->toList(),
+            'TrailingGlobDirs must suffix each directory with /**',
+        );
+    }
+}
